# aws/credentials: Update Credentials cache to use RWMutex